TST: remove enforcement of save order in test_client::test_snap
os.walk is not guaranteed to be deterministic, and test_snap was
intermittently failing in the GitHub test workflow because it assumed
that the last Parameter to be saved to the backend would be the last
Setpoint in the snapshot. Removing this implicit requirement guarantees
the test will pass, and deterministic save order is not a feature
requirement.