MNT: apply PV browser table search every edit
Use QLineEdit.textEdited rather than QLineEdit.editingFinished; this
emits the signal without requiring the user to de-focus the widget,
which feels more natural. If performance is unsatisfactory, we can
adjust all of the filters to reduce computational load.

This change makes the PV browser search bar consistent with the
snapshot table search bar.